+++ packages/known-chains/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Changed individual exports from `default` to `chainSpec`
+
 ## 0.0.1 - 2024-04-10
 
 ### Changed